requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
pytesseract
python-multipart